requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.2.5
django-appconf==1.0.3
django-celery==3.3.1
django-compressor==2.3
django-cors-headers==3.1.0
djangorestframework==3.10.3
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
kombu==3.0.37
Markdown==3.1.1
monotonic==1.5
Pillow==6.1.0
Pygments==2.4.2
pytz==2019.2
rcssmin==1.0.6
redis==2.10.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
